add in the ability to save the model
