+++ paper_reproductions/reinforce/algorithm_implementation.py
@@ -4,8 +4,9 @@ from ..library import DenseLayer, create_dense_neural_net
 
 class Reinforce:
     # environment should have a way to get the state, rewards and say if you are done or not
-    def __init__(self, environment):
+    def __init__(self, environment, save_location):
         self.environment = environment
+        self.save_location = save_location
 
         self.state_size = self.environment.get_state_shape()
         self.action_size = self.environment.get_number_of_actions()
@@ -25,6 +26,8 @@ class Reinforce:
         self.session = tf.Session()
         self.session.run(tf.global_variables_initializer())
 
+        self.saver = tf.train.Saver(tf.global_variables())
+
 
     def _create_learning_network(self, output_units):
 
@@ -40,12 +43,14 @@ class Reinforce:
 
         self.actions_to_take = tf.distributions.Categorical(probs=self.softmax).sample()
 
+        # transform actions taken into one array where one designates the action to take
         self.one_hot_actions = tf.one_hot(self.actions_taken, self.action_size)
+        # get the negative log of the softmax probabilities and then select the one for the action
+        # taken as designated by one_hot_actions
         self.base_cross_entropy = -tf.reduce_sum(tf.math.log(self.softmax + 1e-5) * self.one_hot_actions, axis=1)
         self.reward_weighted_cross_entropy = self.discounted_episode_rewards * self.base_cross_entropy
 
         self.loss = tf.reduce_sum(self.reward_weighted_cross_entropy)
-
         self.train = tf.train.AdamOptimizer(self.learning_rate).minimize(self.loss)
 
     def _discount_rewards(self, non_discounted_rewards):
@@ -58,6 +63,12 @@ class Reinforce:
         return discounted_rewards
 
     def train_model(self, number_of_batches, batch_size, max_steps):
+        try:
+            self.saver.restore(self.session, self.save_location)
+            print('successfully loaded old model data')
+        except:
+            print('do not currently have data stored for this model')
+
         for batch_index in range(number_of_batches):
 
             states = []
@@ -90,19 +101,20 @@ class Reinforce:
             for batch_reward in batch_rewards:
                 discounted_rewards += self._discount_rewards(batch_reward)
 
-            # print('actions taken', actions_taken)
             bce, loss,train = self.session.run([self.base_cross_entropy, self.loss, self.train], feed_dict={
                                                                             self.input:states,
                                                                             self.actions_taken:actions_taken,
                                                                             self.discounted_episode_rewards: discounted_rewards})
 
-            if ((batch_index + 1) % 100) == 0:
+            if ((batch_index + 1) % 10) == 0:
+                self.saver.save(self.session, self.save_location)
                 print('Training Batch: ', batch_index)
-                # print('Base Cross Entroy: ', bce)
-                # print('Final Rewards: ', final_rewards)
                 print('Loss: ', loss)
                 print('Average Number of Steps: ', sum(batch_steps)/len(batch_steps))
                 print('\n')
+
+
+        self.saver.save(self.session, self.save_location)
