+++ paper_reproductions/trainer.py
@@ -2,6 +2,6 @@ from .reinforce.algorithm_implementation import Reinforce
 from .reinforce.environments.cart_pole import CartPole
 
 cart_pole = CartPole(True)
-reinforce = Reinforce(cart_pole)
+reinforce = Reinforce(cart_pole, './.saved_data/cart_pole')
 
-reinforce.train_model(1000, 5, 1000)
+reinforce.train_model(1000, 5, 5000)
